+++ apps/learn-card-app/src/components/new-ai-session/NewAiSessionButton/NewAiSessionButton.tsx
@@ -27,6 +27,7 @@ export enum NewAiSessionButtonEnum {
     revisit,
     default,
     mini,
+    icon,
     mobile,
 }
 
@@ -38,7 +39,7 @@ export const NewAiSessionButton: React.FC<{
     onClick?: () => void;
     className?: string;
     iconType?: 'dark' | 'light';
-}> = ({ type, shortCircuitStep, selectedApp, onClick, className, iconType = 'dark' }) => {
+}> = ({ type, shortCircuitStep, selectedApp, text, onClick, className, iconType = 'dark' }) => {
     const { colors } = useTheme();
     const primaryColor = colors?.defaults?.primaryColor;
 
@@ -157,6 +158,27 @@ export const NewAiSessionButton: React.FC<{
                 <NewAiSessionIcon version="2" className="text-grayscale-900 w-[35px] h-auto" />
             </button>
         );
+    } else if (type === NewAiSessionButtonEnum.icon) {
+        return (
+            <button
+                onClick={
+                    onClick
+                        ? async e => {
+                              e.stopPropagation();
+                              const { prompted } = await gate();
+                              if (prompted) return;
+                              onClick();
+                          }
+                        : e => {
+                              e.stopPropagation();
+                              handleNewSession(undefined, NewAiSessionStepEnum.newTopic);
+                          }
+                }
+                className={`text-[17px] font-semibold font-notoSans text-blue-950 leading-6 rounded-[15px] border-[1px] border-solid !bg-white border-grayscale-200 p-2 flex items-center justify-start mt-[10px] gap-1 shadow-md ${className}`}
+            >
+                <NewAiSessionIcon className="w-[32px] h-[32px]" version={iconType} />
+            </button>
+        );
     } else if (type === NewAiSessionButtonEnum.mobile) {
         return (
             <button
@@ -176,7 +198,7 @@ export const NewAiSessionButton: React.FC<{
                 className={`text-[17px] font-semibold font-notoSans text-blue-950 leading-6 rounded-[15px] border-[1px] border-solid !bg-grayscale-200 border-grayscale-200 p-[10px] w-full flex items-center justify-start mt-[10px] gap-1 ${className}`}
             >
                 {iconType && <NewAiSessionIcon version={iconType} />}
-                New Session
+                {text ?? 'New Session'}
             </button>
         );
     }